finshed admin

+++ server/Controllers/appointments.js
@@ -1,11 +1,11 @@
 const { StatusCodes } = require('http-status-codes')
 const UserAppointment = require('../Database/models/Appointment')
-const  { CustomAPIError }  = require('../errors')
+const { CustomAPIError } = require('../errors')
 
 const GetAllAppointments = async (req, res) => {
     try {
         const appointments = await UserAppointment.find()
-        res.status(StatusCodes.OK).json({requestedUser: req.user, appointments})
+        res.status(StatusCodes.OK).json({ requestedUser: req.user, appointments })
 
     } catch (error) {
         console.log(error);
@@ -16,7 +16,7 @@ const GetAllAppointments = async (req, res) => {
 const GetSpecifiAppointment = async (req, res) => {
     try {
         const appointment = await UserAppointment.findById(req.params.id);
-        res.status(StatusCodes.OK).json({requestedUser: req.user, appointment})
+        res.status(StatusCodes.OK).json({ requestedUser: req.user, appointment })
 
     } catch (error) {
         console.log(error);
@@ -24,35 +24,50 @@ const GetSpecifiAppointment = async (req, res) => {
     }
 }
 const CreateAppointment = async (req, res) => {
-    try {
-        const appointment = await UserAppointment.create({...req.body})
-        res.status(StatusCodes.OK).json({requestedUser: req.user, appointment})
+    if (req.user.role === 'admin') {
+        try {
+            const appointment = await UserAppointment.create({ ...req.body })
+            res.status(StatusCodes.OK).json({ requestedUser: req.user, appointment })
 
-    } catch (error) {
-        console.log(error);
-        throw new CustomAPIError('Internal Server Error')
+        } catch (error) {
+            console.log(error);
+            throw new CustomAPIError('Internal Server Error')
+        }
+    }
+    else {
+        res.status(StatusCodes.BAD_REQUEST).json({ role: req.user.role, message: 'Unauthorized access' })
     }
 }
 
 const UpdateAppointment = async (req, res) => {
-    try {
-        const appointment = await UserAppointment.findByIdAndUpdate(req.params.id, req.body, {new: true});
-        res.status(StatusCodes.OK).json({requestedUser: req.user, appointment})
+    if (req.user.role === 'admin') {
+        try {
+            const appointment = await UserAppointment.findByIdAndUpdate(req.params.id, req.body, { new: true });
+            res.status(StatusCodes.OK).json({ requestedUser: req.user, appointment })
 
-    } catch (error) {
-        console.log(error);
-        throw new CustomAPIError('Internal Server Error')
+        } catch (error) {
+            console.log(error);
+            throw new CustomAPIError('Internal Server Error')
+        }
+    }
+    else {
+        res.status(StatusCodes.BAD_REQUEST).json({ role: req.user.role, message: 'Unauthorized access' })
     }
 }
 
 const DeletAppointment = async (req, res) => {
-    try {
-        const appointment = await UserAppointment.findByIdAndDelete(req.params.id);
-        res.status(StatusCodes.OK).json({requestedUser: req.user, appointment})
+    if (req.user.role === 'admin') {
+        try {
+            const appointment = await UserAppointment.findByIdAndDelete(req.params.id);
+            res.status(StatusCodes.OK).json({ requestedUser: req.user, appointment })
 
-    } catch (error) {
-        console.log(error);
-        throw new CustomAPIError('Internal Server Error')
+        } catch (error) {
+            console.log(error);
+            throw new CustomAPIError('Internal Server Error')
+        }
+    }
+    else {
+        res.status(StatusCodes.BAD_REQUEST).json({ role: req.user.role, message: 'Unauthorized access' })
     }
 }
